+++ gameservice/src/main/java/jl/gameservice/persistence/GameRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GameRepository extends JpaRepository<GameEntity, Integer> {
+public interface GameRepository extends JpaRepository<GameEntity, Long> {
 
 }
